requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-experimental
duckduckgo-search
pandas
requests
python-dotenv
streamlit
openai
faiss-cpu
sentence-transformers
langchain-groq
